+++ ubo_app/services/080-docker/ubo_handle.py
@@ -11,9 +11,9 @@ def setup(service: Service) -> None:
     from reducer import reducer
     from setup import init_service
 
-    service.register_reducer(reducer)
-
     init_service()
+
+    service.register_reducer(reducer)
 
 
 register(